% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd_inStock.R
\name{prd_inStock_log_view}
\alias{prd_inStock_log_view}
\title{日志查询}
\usage{
prd_inStock_log_view(dms_token, FSrcBillNo, FDate)
}
\arguments{
\item{FEndDate}{}
}
\value{

}
\description{
日志查询
}
\examples{
prd_inStock_log_view()
}
