% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd_pickMtrl2.R
\name{prd_pickMtrl2_Fisdo_update}
\alias{prd_pickMtrl2_Fisdo_update}
\title{更新同步状态}
\usage{
prd_pickMtrl2_Fisdo_update(dms_token, FSrcBillNo, FDate)
}
\arguments{
\item{FEndDate}{}
}
\value{

}
\description{
更新同步状态
}
\examples{
prd_pickMtrl2_Fisdo_update()
}
