% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pur_inStock.R
\name{pur_inStock_log_view}
\alias{pur_inStock_log_view}
\title{日志查询}
\usage{
pur_inStock_log_view(dms_token, FSrcBillNo, FDate)
}
\arguments{
\item{FEndDate}{}
}
\value{

}
\description{
日志查询
}
\examples{
pur_inStock_log_view()
}
