% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pur_returnStock.R
\name{pur_returnStock_Fisdo_update}
\alias{pur_returnStock_Fisdo_update}
\title{更新同步状态}
\usage{
pur_returnStock_Fisdo_update(dms_token, FSrcBillNo, FDate)
}
\arguments{
\item{FEndDate}{}
}
\value{

}
\description{
更新同步状态
}
\examples{
pur_returnStock_Fisdo_update()
}
