% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_returnStock.R
\name{sal_returnStock_Fisdo_update}
\alias{sal_returnStock_Fisdo_update}
\title{更新同步状态}
\usage{
sal_returnStock_Fisdo_update(dms_token, FSrcBillNo, FDate)
}
\arguments{
\item{FEndDate}{}
}
\value{

}
\description{
更新同步状态
}
\examples{
sal_returnStock_Fisdo_update()
}
