% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_returnStock.R
\name{sal_returnStock_log_view}
\alias{sal_returnStock_log_view}
\title{日志查询}
\usage{
sal_returnStock_log_view(dms_token, FSrcBillNo, FDate)
}
\arguments{
\item{FEndDate}{}
}
\value{

}
\description{
日志查询
}
\examples{
sal_returnStock_log_view()
}
