% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_AssembledApp.R
\name{stk_AssembledApp_Fisdo_update}
\alias{stk_AssembledApp_Fisdo_update}
\title{更新同步状态}
\usage{
stk_AssembledApp_Fisdo_update(dms_token, FSrcBillNo, FDate)
}
\arguments{
\item{FEndDate}{}
}
\value{

}
\description{
更新同步状态
}
\examples{
stk_AssembledApp_Fisdo_update()
}
