% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_transferDirect.R
\name{stk_transferDirect_Fisdo_update}
\alias{stk_transferDirect_Fisdo_update}
\title{更新同步状态}
\usage{
stk_transferDirect_Fisdo_update(dms_token, FBillTypeID, FINDEPTID, FDate)
}
\arguments{
\item{FEndDate}{}
}
\value{

}
\description{
更新同步状态
}
\examples{
stk_transferDirect_Fisdo_update()
}
