% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_transferDirect.R
\name{stk_transferDirect_log_view}
\alias{stk_transferDirect_log_view}
\title{日志查询}
\usage{
stk_transferDirect_log_view(dms_token, FBillTypeID, FINDEPTID, FDate)
}
\arguments{
\item{FEndDate}{}
}
\value{

}
\description{
日志查询
}
\examples{
stk_transferDirect_log_view()
}
