% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_pickMtrl.R
\name{sub_pickMtrl_log_view}
\alias{sub_pickMtrl_log_view}
\title{日志查询}
\usage{
sub_pickMtrl_log_view(dms_token, FSrcBillNo, FDate)
}
\arguments{
\item{FEndDate}{}
}
\value{

}
\description{
日志查询
}
\examples{
sub_pickMtrl_log_view()
}
