% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_returnMtrl.R
\name{sub_returnMtrl_Fisdo_update}
\alias{sub_returnMtrl_Fisdo_update}
\title{更新同步状态}
\usage{
sub_returnMtrl_Fisdo_update(dms_token, FSrcBillNo, FDate)
}
\arguments{
\item{FEndDate}{}
}
\value{

}
\description{
更新同步状态
}
\examples{
sub_returnMtrl_Fisdo_update()
}
