% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_miscellaneous.R
\name{stk_miscellaneous_Fisdo_update}
\alias{stk_miscellaneous_Fisdo_update}
\title{更新同步状态}
\usage{
stk_miscellaneous_Fisdo_update(dms_token, FBillTypeID, FDEPTID, FDate)
}
\arguments{
\item{dms_token}{数据中台}

\item{FBillTypeID}{单据类型}

\item{FDEPTID}{部门代码}

\item{FDate}{日期}
}
\value{

}
\description{
更新同步状态
}
\examples{
stk_miscellaneous_Fisdo_update()
}
