% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_miscellaneous.R
\name{stk_miscellaneous_log_view}
\alias{stk_miscellaneous_log_view}
\title{日志查询}
\usage{
stk_miscellaneous_log_view(dms_token, FBillTypeID, FDEPTID, FDate)
}
\arguments{
\item{FEndDate}{}
}
\value{

}
\description{
日志查询
}
\examples{
stk_miscellaneous_log_view()
}
